backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
pydantic-settings>=2.4,<3.0
python-multipart>=0.0.9,<1.0
pytest>=8.0,<9.0
